Modifications has been made on this files to suupport Hops distributed YARN

+++ src/main/java/io/hops/metadata/yarn/dal/PendingEventDataAccess.java
@@ -43,7 +43,7 @@ public interface PendingEventDataAccess<T> extends EntityDataAccess {
    * @return
    * @throws StorageException
    */
-  List<T> getAll(byte status) throws StorageException;
+  List<T> getAll(int status) throws StorageException;
 
   /**
    * @param pendingEvent
